+++ crates/rust-client/src/notes/note_screener.rs
@@ -87,8 +87,13 @@ impl NoteScreener {
             "Should be able to convert the first two note inputs to an array of two Felt elements",
         );
 
-        let account_id = AccountId::try_from([account_id_felts[1], account_id_felts[0]])
-            .map_err(|err| InvalidNoteInputsError::AccountError(note.id(), AccountError::FinalAccountHeaderIdParsingFailed(err),))?;
+        let account_id =
+            AccountId::try_from([account_id_felts[1], account_id_felts[0]]).map_err(|err| {
+                InvalidNoteInputsError::AccountError(
+                    note.id(),
+                    AccountError::FinalAccountHeaderIdParsingFailed(err),
+                )
+            })?;
 
         if !account_ids.contains(&account_id) {
             return Ok(vec![]);
@@ -116,8 +121,13 @@ impl NoteScreener {
             InvalidNoteInputsError::BlockNumberError(note.id(), recall_height_felt.as_int())
         })?;
 
-        let account_id = AccountId::try_from([account_id_felts[1], account_id_felts[0]])
-            .map_err(|err| InvalidNoteInputsError::AccountError(note.id(), AccountError::FinalAccountHeaderIdParsingFailed(err)))?;
+        let account_id =
+            AccountId::try_from([account_id_felts[1], account_id_felts[0]]).map_err(|err| {
+                InvalidNoteInputsError::AccountError(
+                    note.id(),
+                    AccountError::FinalAccountHeaderIdParsingFailed(err),
+                )
+            })?;
 
         Ok(vec![
             (account_id, NoteRelevance::Always),