+++ crates/rust-client/src/store/sqlite_store/accounts.rs
@@ -5,7 +5,9 @@ use alloc::{
 use std::{collections::BTreeMap, rc::Rc};
 
 use miden_objects::{
-    accounts::{Account, AccountCode, AccountHeader, AccountId, AccountStorage, AuthSecretKey}, assets::{Asset, AssetVault}, AccountError, Digest, Felt, Word
+    accounts::{Account, AccountCode, AccountHeader, AccountId, AccountStorage, AuthSecretKey},
+    assets::{Asset, AssetVault},
+    AccountError, Digest, Felt, Word,
 };
 use miden_tx::utils::{Deserializable, Serializable};
 use rusqlite::{params, types::Value, Connection, Transaction};
@@ -213,7 +215,11 @@ impl SqliteStore {
                 result.map_err(|err| StoreError::ParsingError(err.to_string())).and_then(
                     |(id, code): (String, Vec<u8>)| {
                         Ok((
-                            AccountId::from_hex(&id).map_err(|err| StoreError::AccountError(AccountError::FinalAccountHeaderIdParsingFailed(err)))?,
+                            AccountId::from_hex(&id).map_err(|err| {
+                                StoreError::AccountError(
+                                    AccountError::FinalAccountHeaderIdParsingFailed(err),
+                                )
+                            })?,
                             AccountCode::from_bytes(&code).map_err(StoreError::AccountError)?,
                         ))
                     },